Continue developing DroidTracer as a module
DroidTracer was developed since 2013 as part of the MonitorMe app
monolithically. Now it was extracted, and will be maintained
separately as an Android Studio module; i.e, it is here to be enjoyed
by everyone to develop their own app where intercepting the
behaviour of other running apps in needed.

It will be used in MonitorMe from now on as a submodule.

Things I might add to DroidTracer in the future are in the TODO file.