Send droidtracer UID from native code to kernel
Remove setDroidTracerUid(int uid) from API in DroidTracerService.java,
and instead send the UID of droidtracer from droidtracer.cpp, during the
initialisation of netlink, to the kernel module. The reason is that we _never_
want to trace droidtracer itself, so why not automatically tell the kernel
the UID already during netlink init phase.